requirements.txt
# Reddit API (async)
asyncpraw>=7.7.0

# RSS Feeds
feedparser>=6.0.0
aiohttp>=3.9.0

# Database
aiosqlite<=0.17.0

# Discord
discord-webhook>=1.3.0    # For webhook mode
discord.py>=2.3.0         # For bot mode (multi-server)

# NLP
spacy>=3.7.0
scikit-learn>=1.4.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Iceout.org API (MessagePack format + headless browser for auth)
msgpack>=1.0.0
playwright>=1.40.0

# Utilities
python-dateutil>=2.8.0
